+++ extras/spytools/src/test/java/arez/spytools/SpyUtilTest.java
@@ -1,6 +1,8 @@
 package arez.spytools;
 
 import arez.spy.ObserverCreateEvent;
+import arez.spy.ReactionCycleCompleteEvent;
+import arez.spy.ReactionCycleStartEvent;
 import arez.spy.TransactionCompleteEvent;
 import arez.spy.TransactionStartEvent;
 import org.testng.annotations.Test;
@@ -15,6 +17,8 @@ public class SpyUtilTest
     assertEquals( SpyUtil.getNestingDelta( Object.class ), SpyUtil.NestingDelta.UNKNOWN );
     assertEquals( SpyUtil.getNestingDelta( TransactionStartEvent.class ), SpyUtil.NestingDelta.INCREASE );
     assertEquals( SpyUtil.getNestingDelta( TransactionCompleteEvent.class ), SpyUtil.NestingDelta.DECREASE );
+    assertEquals( SpyUtil.getNestingDelta( ReactionCycleStartEvent.class ), SpyUtil.NestingDelta.INCREASE );
+    assertEquals( SpyUtil.getNestingDelta( ReactionCycleCompleteEvent.class ), SpyUtil.NestingDelta.DECREASE );
     assertEquals( SpyUtil.getNestingDelta( ObserverCreateEvent.class ), SpyUtil.NestingDelta.NONE );
   }
 }